Move tests to be interleaved with source files
